Split net core framework location functionality into a separate class
